+++ src/test/resources/cucumberTemplateTest.feature
@@ -1,0 +1,6 @@
+Feature: Setup of Cucumber Template
+
+  Scenario: Cucumber test feature runs
+    Given Test feature file is ready
+    When I run the feature file
+    Then Run should be successfully
