--- src/test/resources/cucumber_template_test.feature
@@ -1,6 +1,0 @@
-Feature: Setup of Cucumber Template
-
-  Scenario: Cucumber test feature runs
-    Given Test feature file is ready
-    When I run the feature file
-    Then Run should be successfully
